refactor: CQRS 서비스 구조로 변경
