+++ src/main/java/com/pitchain/service/IndividualQueryService.java
@@ -1,0 +1,21 @@
+package com.pitchain.service;
+
+import com.pitchain.common.constant.OauthProvider;
+import com.pitchain.entity.Individual;
+import com.pitchain.repository.IndividualRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
+
+@RequiredArgsConstructor
+@Service
+public class IndividualQueryService {
+    private final IndividualRepository individualRepository;
+
+    @Transactional(readOnly = true)
+    public Optional<Individual> findByOauthProviderAndSocialId(OauthProvider oauthProvider, String socialId) {
+        return individualRepository.findByOauthProviderAndSocialId(oauthProvider, socialId);
+    }
+}
