+++ src/main/java/com/pitchain/service/MyBmHistoryCommandService.java
@@ -8,13 +8,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@RequiredArgsConstructor
 @Service
-public class MyBmHistoryService {
-    private final MyBmHistoryCommandService myBmHistoryCommandService;
+@RequiredArgsConstructor
+public class MyBmHistoryCommandService {
+    private final MyBmHistoryRepository myBmHistoryRepository;
 
     @Transactional
     public void saveMyBmHistory(Member member, Bm bm) {
-        myBmHistoryCommandService.saveMyBmHistory(member, bm);
+        if (!myBmHistoryRepository.existsByMemberAndBm(member, bm))
+            myBmHistoryRepository.save(MyBmHistory.of(member, bm));
     }
 }
