+++ src/main/java/com/pitchain/service/MySpHistoryService.java
@@ -1,8 +1,5 @@
 package com.pitchain.service;
 
-import com.pitchain.entity.Bm;
-import com.pitchain.entity.Member;
-import com.pitchain.entity.MySpHistory;
 import com.pitchain.jwt.MemberDetails;
 import com.pitchain.repository.EntityFacade;
 import com.pitchain.repository.MySpHistoryRepository;
@@ -13,19 +10,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class MySpHistoryService {
-
-    private final MySpHistoryRepository mySpHistoryRepository;
-    private final EntityFacade entityFacade;
+    private final MySpHistoryCommandService mySpHistoryCommandService;
 
     @Transactional
     public void saveMySpHistory(MemberDetails memberDetails, Long bmId, int viewTime) {
-        Member member = entityFacade.getMember(memberDetails.id());
-        Bm bm = entityFacade.getBm(bmId);
-
-        mySpHistoryRepository.findByMemberAndBm(member, bm)
-                .ifPresentOrElse(
-                        existingHistory -> existingHistory.updateViewTime(viewTime),
-                        () -> mySpHistoryRepository.save(new MySpHistory(member, bm, viewTime))
-                );
+        mySpHistoryCommandService.saveMySpHistory(memberDetails, bmId, viewTime);
     }
 }
