+++ src/main/java/com/pitchain/service/PtImgQueryService.java
@@ -1,6 +1,9 @@
 package com.pitchain.service;
 
+import com.pitchain.entity.Bm;
 import com.pitchain.entity.PtImg;
+import com.pitchain.repository.EntityFacade;
+import com.pitchain.repository.PtImgRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -9,18 +12,14 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @Service
-public class PtImgService {
-    private final PtImgCommandService ptImgCommandService;
-    private final PtImgQueryService ptImgQueryService;
+public class PtImgQueryService {
+    private final EntityFacade entityFacade;
+    private final PtImgRepository ptImgRepository;
 
     @Transactional(readOnly = true)
     public List<PtImg> getPtImgsByBmId(Long bmId) {
-        return ptImgQueryService.getPtImgsByBmId(bmId);
-    }
+        Bm bm = entityFacade.getBm(bmId);
 
-    @Transactional
-    public void update(Long bmId, List<String> uploadPtImgKeys) {
-        ptImgCommandService.updatePtImgs(bmId, uploadPtImgKeys);
+        return ptImgRepository.findAllByBmId(bm.getId());
     }
-
 }
